功能更新，bug修复。

+++ src/main/java/com/xxsword/xitem/admin/domain/course/entity/CourseFileItem.java
@@ -30,6 +30,6 @@ public class CourseFileItem extends BaseEntity implements Serializable {
 
     @Column
     @ColumnComment("排序")
-    private Long seq;
+    private Integer seq;
 
 }
